moved evaluation metrics to separate file
